ai/pyproject.toml
[tool.poetry]
name = "email-summarisation"
version = "0.1.0"
description = ""
authors = ["Oliver Yerbury-Hodgson <oyerburyhodgson@gmail.com>"]
readme = "README.md"
packages = [{include = "email_summarisation"}]

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^4.0.9"
langchain = "^0.0.350"
pandas = "^2.1.4"
pytest = "^7.4.3"
spacy = "^3.7.2"
names-dataset = "^3.1.0"
tqdm = "^4.66.1"
python-dotenv = "^1.0.0"
openai = "^1.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
